+++ pkg/models/query.go
@@ -115,6 +115,7 @@ type Query struct {
 	DataOverrides                      []InfinityDataOverride `json:"dataOverrides"`
 	GlobalQueryID                      string                 `json:"global_query_id"`
 	QueryMode                          string                 `json:"query_mode"`
+	Method                             string                 `json:"method,omitempty"`
 	Spreadsheet                        string                 `json:"spreadsheet,omitempty"`
 	SheetName                          string                 `json:"sheetName,omitempty"`
 	SheetRange                         string                 `json:"sheetRange,omitempty"`
@@ -214,8 +215,14 @@ func ApplyDefaultsToQuery(ctx context.Context, pCtx *backend.PluginContext, quer
 			query.URL = "https://raw.githubusercontent.com/grafana/grafana-infinity-datasource/main/testdata/users.json"
 		}
 	}
-	if query.Source == "url" && strings.TrimSpace(query.URLOptions.Method) == "" {
-		query.URLOptions.Method = http.MethodGet
+	if query.Source == "url" {
+		query.URLOptions.Method = strings.TrimSpace(query.URLOptions.Method)
+		if query.URLOptions.Method == "" {
+			query.URLOptions.Method = strings.TrimSpace(query.Method)
+			if query.URLOptions.Method == "" {
+				query.URLOptions.Method = http.MethodGet
+			}
+		}
 	}
 	if query.Source == "url" && (!strings.EqualFold(query.URLOptions.Method, http.MethodGet)) {
 		if query.URLOptions.BodyType == "" {